引入redission 流水号模块优化

+++ sc/sequence-service/src/main/java/cloud/flystar/solon/sequence/service/bo/Segment.java
@@ -25,7 +25,7 @@ public class Segment {
     /**
      * 步长
      */
-    public volatile int step;
+    public volatile long step;
 
 
     /**
